Incluindo suporte a leitura de grafos no formato Matrix Market (.mtx). Corrigindo bug que causava erro na leitura de grafos com numero de vertices maior que o limite da variavel int.

+++ graph-common/src/graph/include/GraphDefs.h
@@ -26,12 +26,12 @@ struct Edge {
 	}
 };
 struct Vertex {
-    int id;
+    long id;
     std::size_t edge_index_t;
     std::size_t vertex_rank_t;
 
     Vertex() : id(0), edge_index_t(0) { }
-    Vertex(int w) : id(w), edge_index_t(0), vertex_rank_t(0)  { }
+    Vertex(long w) : id(w), edge_index_t(0), vertex_rank_t(0)  { }
 
     friend class boost::serialization::access;
